secp256k1: Move 4x64 field code to subpackage.
This moves the entire FieldVal64 implementation, tests, and benchmarks
into a separate package named field4x64 and only makes the minimal
required changes to ensure the code compiles and all tests pass.

It also modifies the Normalize method to match the signature used by
FieldVal for API parity.

No functional changes are intended by this commit.  It is strictly a
refactoring that preserves the existing behavior.

This is part of a series of commits that will ultimately allow
architecture-specific backend selection.
