+++ dcrec/secp256k1/field4x64/field64_amd64_test.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
 
-package secp256k1
+package field4x64
 
 import "testing"
