+++ dcrec/secp256k1/field4x64/field64_bench_amd64_test.go
@@ -4,7 +4,7 @@
 
 //go:build amd64 && !purego
 
-package secp256k1
+package field4x64
 
 import (
 	"testing"
