+++ dcrec/secp256k1/field4x64/field64_noasm.go
@@ -4,7 +4,7 @@
 
 //go:build !amd64 || purego
 
-package secp256k1
+package field4x64
 
 // field64UseBMI2AndADX depends on hardware support and access to assembly
 // instructions.
